requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
certifi==2025.10.5
cffi==2.0.0
click==8.1.8
cryptography==46.0.3
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.119.0
h11==0.16.0
httptools==0.7.1
idna==3.11
influxdb3-python==0.16.0
passlib==1.7.4
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
reactivex==4.0.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
